+++ packages/evokit-box/CHANGELOG.md
@@ -1,9 +1,9 @@
 [README]: /packages/evokit-box/README.md
-
 [box-margin]: /packages/evokit-box/?id=box-margin
 [box-overflow]: /packages/evokit-box/?id=box-overflow
 [box-padding]: /packages/evokit-box/?id=box-padding
 [box-round]: /packages/evokit-box/?id=box-round
+
 [create_theme]: /docs/base/theme.md
 
 # EvoKit - Box
@@ -27,7 +27,7 @@
     - `box-margin-tb`
     - `box-padding-lr`
     - `box-padding-tb`
-    - `box-round-top`,
+    - `box-round-top`
     - `box-round-right`
     - `box-round-bottom`
     - `box-round-left`